修改包位置，增加异常处理，优化模块分工

+++ clever-framework-web/src/main/java/io/github/toquery/framework/web/domain/ResponsePage.java
@@ -15,11 +15,11 @@ import lombok.Setter;
 @AllArgsConstructor
 public class ResponsePage {
 
-    private Integer pageSize;
+    private int pageSize;
 
-    private Integer pageNumber;
+    private int pageNumber;
 
-    private Long totalElements;
+    private int totalElements;
 
-    private Integer totalPages;
+    private int totalPages;
 }
